requirements.txt
streamlit
torch
pillow
pandas
spotipy
git+https://github.com/openai/CLIP.git
